feat :: ReviewDetailsScreenViewModel
